# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * QC Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  database: string;
  scheduler: string;
  api: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  Scheduled: 'Scheduled',
  Upcoming: 'Upcoming',
  Due: 'Due',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  scheduledDate: string;
  scheduledTime: string;
  scheduledAt: string;
  status: TaskStatus;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  completedAt: string | null;
}

export interface TaskInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /** @maxLength 500 */
  description?: string;
  scheduledDate: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  scheduledTime: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  Scheduled: 'Scheduled',
  Upcoming: 'Upcoming',
  Due: 'Due',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface TaskUpdate {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title?: string;
  /**
     * @maxLength 500
     * @nullable
     */
  description?: string | null;
  scheduledDate?: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  scheduledTime?: string;
  status?: TaskUpdateStatus;
}

export interface DashboardStats {
  total: number;
  scheduled: number;
  upcoming: number;
  due: number;
  completed: number;
  cancelled: number;
  nextTask: Task | null;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  created: 'created',
  updated: 'updated',
  completed: 'completed',
  due: 'due',
  rescheduled: 'rescheduled',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  message: string;
  timestamp: string;
  task: Task;
}

export type ListTasksParams = {
status?: ListTasksStatus;
/**
 * @maxLength 100
 */
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  All: 'All',
  Scheduled: 'Scheduled',
  Upcoming: 'Upcoming',
  Due: 'Due',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export type GetRecentActivityParams = {
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};