# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QC Task Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management operations
  - name: dashboard
    description: Dashboard summaries and activity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [All, Scheduled, Upcoming, Due, Completed, Cancelled]
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid task input
  /tasks/due:
    get:
      operationId: getDueTasks
      tags: [tasks]
      summary: Get tasks that need a reminder
      responses:
        "200":
          description: Due tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update or reschedule a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid task input
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task completed
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard counts
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent task activity
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        database:
          type: string
        scheduler:
          type: string
        api:
          type: string
      required:
        - status
        - database
        - scheduler
        - api
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [Scheduled, Upcoming, Due, Completed, Cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - title
        - description
        - scheduledDate
        - scheduledTime
        - scheduledAt
        - status
        - createdAt
        - updatedAt
        - completedAt
    TaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 500
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
      required:
        - title
        - scheduledDate
        - scheduledTime
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: ["string", "null"]
          maxLength: 500
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        status:
          type: string
          enum: [Scheduled, Upcoming, Due, Completed, Cancelled]
    DashboardStats:
      type: object
      properties:
        total:
          type: integer
        scheduled:
          type: integer
        upcoming:
          type: integer
        due:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
        nextTask:
          anyOf:
            - $ref: "#/components/schemas/Task"
            - type: "null"
      required:
        - total
        - scheduled
        - upcoming
        - due
        - completed
        - cancelled
        - nextTask
    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [created, updated, completed, due, rescheduled]
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        task:
          $ref: "#/components/schemas/Task"
      required:
        - id
        - type
        - message
        - timestamp
        - task