# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * QC Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityType } from './activityType';
import type { Task } from './task';

export interface Activity {
  id: string;
  type: ActivityType;
  message: string;
  timestamp: Date;
  task: Task;
}