# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * QC Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';

export interface DashboardStats {
  total: number;
  scheduled: number;
  upcoming: number;
  due: number;
  completed: number;
  cancelled: number;
  nextTask: Task | null;
}