# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * QC Task Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
  database: string;
  scheduler: string;
  api: string;
}