# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * QC Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  scheduledDate: Date;
  scheduledTime: string;
  scheduledAt: Date;
  status: TaskStatus;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  completedAt: Date | null;
}