# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * QC Task Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /** @maxLength 500 */
  description?: string;
  scheduledDate: Date;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  scheduledTime: string;
}